+++ lib/DAV/CalendarObject.php
@@ -1,0 +1,115 @@
+<?php
+
+namespace OCA\SalatTime\DAV;
+
+use Sabre\VObject\Component\VCalendar;
+
+class CalendarObject implements \Sabre\CalDAV\ICalendarObject, \Sabre\DAVACL\IACL {
+	/** @var Calendar */
+	private $calendar;
+	/** @var string */
+	private $name;
+	/** @var VCalendar */
+	private $calendarObject;
+	/**
+	 * CalendarObject constructor.
+	 *
+	 * @param Calendar $calendar
+	 * @param string $name
+	 * @param CalculationService $calculationService
+	 */
+	public function __construct(Calendar $calendar, string $name) {
+		$this->calendar = $calendar;
+		$this->name = $name;
+		$this->calendarObject = $this->getCalendarObject();
+	}
+
+	public function getOwner() {
+		return null;
+	}
+
+	public function getGroup() {
+		return null;
+	}
+
+	public function getACL() {
+		return $this->calendar->getACL();
+	}
+
+	public function setACL(array $acl) {
+		throw new \Sabre\DAV\Exception\Forbidden('Setting ACL is not supported on this node');
+	}
+
+	public function getSupportedPrivilegeSet() {
+		return null;
+	}
+
+	public function put($data) {
+		throw new \Sabre\DAV\Exception\Forbidden('This calendar-object is read-only');
+	}
+
+	public function get() {
+		if ($this->calendarObject) {
+			return $this->calendarObject->serialize();
+		}
+	}
+
+	public function getContentType() {
+		return 'text/calendar; charset=utf-8';
+	}
+
+	public function getETag() {
+		return '"' . md5($this->get()) . '"';
+		////return '"' . md5($this->sourceItem->getLastModified()) . '"';
+	}
+
+	public function getSize() {
+		return mb_strlen($this->calendarObject->serialize());
+	}
+
+	public function delete() {
+		throw new \Sabre\DAV\Exception\Forbidden('This calendar-object is read-only');
+	}
+
+	public function getName() {
+		return $this->name;
+	}
+
+	public function setName($name) {
+		throw new \Sabre\DAV\Exception\Forbidden('This calendar-object is read-only');
+	}
+
+	public function getLastModified() {
+		return time();
+		///            return $this->sourceItem->getLastModified();
+	}
+
+	private function getCalendarObject(): VCalendar {
+		$calendar = new VCalendar();
+		$name = $this->getName();
+		$Data = $this->extractData($name);
+		if ($Data) {
+			$eDate = $this->calendar->getEventDate($Data[1], $Data[0]);
+			$event = $calendar->createComponent('VEVENT');
+			$event->UID = $name;
+			$event->DTSTAMP = $eDate;   //gmdate('Ymd\\THis\\Z');
+			$event->DTSTART = $eDate;
+			$event->DURATION = "PT10M";
+			$event->SUMMARY = $this->calendar->getEventSummary($Data[1], $Data[0]);
+			$event->DESCRIPTION = $this->calendar->getEventDescription($Data[1], $Data[0]);
+
+			$calendar->add($event);
+		}
+
+		return $calendar;
+	}
+
+	private function extractData($name) {
+		$parts = explode('_', substr($name, 0, -4));
+		if (count($parts) === 2) {
+			return $parts;
+		}
+
+		return null;
+	}
+}