+++ templates/settings/index.php
@@ -5,8 +5,9 @@
 		></button>
 	</div>
 	<div id="app-settings-content">
-		<!-- Your settings in here -->
                 <input type="checkbox" id="salatnotification" class="checkbox" name="salatnotification" <?php if ($_['notification'] == 'true') echo "checked"; ?> >
                 <label for="salatnotification"> <?php echo $l->t('Enable notification'); ?> </label>
+                <input type="checkbox" id="salatcalendar" class="checkbox" name="salatcalendar" <?php if ($_['calendar'] == 'true') echo "checked"; ?> >
+                <label for="salatcalendar"> <?php echo $l->t('Enable Calendar'); ?> </label>
 	</div>
 </div>